test/ani_namespace: fix the problem that i8 was output as a character, resulting in illegal characters

+++ test/ani_namespace/author/src/inner.impl.cpp
@@ -96,7 +96,7 @@ public:
   }
 
   int8_t geti8() {
-    std::cout << __func__ << " " << i8 << std::endl;
+    std::cout << __func__ << " " << (int)i8 << std::endl;
     return i8;
   }
